requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
idna==2.9
pathspec==0.8.0
Pillow==7.1.2
python-dotenv==0.13.0
regex==2020.4.4
requests==2.23.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
Wand==0.5.9
